+++ packages/core/src/security/index.ts
@@ -4,3 +4,4 @@
 
 export * from './checks.js';
 export * from './anomaly-rules.js';
+export * from './firewall.js';